style: 커뮤니티게시판(디테일페이지, 수정/등록페이지)퍼블리싱
