+++ src/app/communityBoard/detail/page.tsx
@@ -1,0 +1,8 @@
+import CommunityBoardDetail from "@/components/CommunityBoard-detail";
+import React from "react";
+
+const page = () => {
+  return <CommunityBoardDetail />;
+};
+
+export default page;
