+++ src/app/communityBoard/page.tsx
@@ -1,0 +1,8 @@
+import CommunityBoardNew from "@/components/CommunityBoard-new";
+import React from "react";
+
+const page = () => {
+  return <CommunityBoardNew />;
+};
+
+export default page;
