+++ src/app/jobList/[boardId]/edit/page.tsx
@@ -1,4 +1,4 @@
-import JobBoardWritePage from "@/components/JobBoard-write";
+import JobBoardWritePage from "@/components/JobBoard-new";
 import React from "react";
 
 const JobListEdit = () => {
