+++ src/components/CommunityBoard-new/index.tsx
@@ -1,0 +1,60 @@
+"use client";
+import Image from "next/image";
+import Button from "@/commons/Button";
+import Input from "@/commons/input";
+
+const CommunityBoardNew = () => {
+  return (
+    <div className="min-h-screen flex flex-col">
+      <form className="p-4 space-y-6 flex-1">
+        <div>
+          <label className="block text-sm text-text-primary mb-1">제목</label>
+          <Input
+            type="text"
+            placeholder="제목을 입력해주세요"
+            className="w-full"
+          />
+        </div>
+
+        <div>
+          <label className="block text-sm text-text-primary mb-1">
+            상세 내용
+          </label>
+          <textarea
+            placeholder="내용을 입력해주세요"
+            className="resize-none flex w-full h-[13rem] px-4 py-4 items-center gap-2 self-stretch rounded-xl border focus:border-[rgba(27,141,90,0.93)] focus:outline-none"
+          />
+        </div>
+
+        <div>
+          <label className="block text-sm font-medium text-text-primary mb-2">
+            사진 첨부
+          </label>
+          <div className="flex gap-4 flex-wrap">
+            <div className="w-[100px] h-[100px] border-2 border-dashed border-gray-300 rounded-lg flex items-center justify-center cursor-pointer hover:border-gray-400 transition-colors">
+              <Image
+                src="/icons/camera.png"
+                alt="Upload Image"
+                width={40}
+                height={40}
+              />
+            </div>
+          </div>
+        </div>
+
+        <div className="w-full">
+          <Button
+            type="submit"
+            design="design1"
+            width="full"
+            className="h-[3.5rem]"
+          >
+            등록하기
+          </Button>
+        </div>
+      </form>
+    </div>
+  );
+};
+
+export default CommunityBoardNew;
