+++ src/components/JobBoard-new/index.tsx
@@ -1,0 +1,221 @@
+"use client";
+import { Controller } from "react-hook-form";
+import { z } from "zod";
+import Button from "@/commons/Button";
+import Input from "@/commons/input";
+import Image from "next/image";
+import RegionDropdown from "@/commons/regionsDropdown";
+import { JobListNewProps } from "./types";
+import { useJobBoardNew } from "./hook";
+
+// 폼 스키마 수정
+export const jobFormSchema = z.object({
+  title: z.string().min(1, "제목을 입력해주세요"),
+  mainRegion: z.string().min(1, "지역을 선택해주세요"),
+  subRegion: z.string().min(1, "세부 지역을 선택해주세요"),
+  price: z
+    .string()
+    .min(1, "금액을 입력해주세요")
+    .regex(/^[0-9,]+$/, "올바른 금액 형식을 입력해주세요"),
+  contents: z.string().min(1, "상세 내용을 입력해주세요"),
+  newImages: z.array(z.instanceof(File)).optional(),
+});
+
+const JobBoardNew = ({ isEdit }: JobListNewProps) => {
+  const {
+    control,
+    handleSubmit,
+    errors,
+    handleOpenFileDialog,
+    handleFileChange,
+    onSubmit,
+    existingImages,
+    handleDeleteExistingImage,
+    watch,
+    setValue,
+    fileInputRef,
+  } = useJobBoardNew();
+
+  return (
+    <div className="min-h-screen flex flex-col">
+      <form onSubmit={handleSubmit(onSubmit)} className="p-4 space-y-6 flex-1">
+        <div>
+          <label className="block text-sm text-text-primary mb-1">제목</label>
+          <Controller
+            name="title"
+            control={control}
+            render={({ field }) => (
+              <Input
+                {...field}
+                type="text"
+                placeholder="제목을 입력해주세요"
+                className="w-full"
+              />
+            )}
+          />
+          {errors.title && (
+            <p className="text-red-500 text-sm mt-1">{errors.title.message}</p>
+          )}
+        </div>
+
+        <div>
+          <label className="block text-sm text-text-primary mb-1">
+            지역 선택
+          </label>
+          <Controller
+            name="mainRegion"
+            control={control}
+            render={({ field: { value, onChange } }) => (
+              <Controller
+                name="subRegion"
+                control={control}
+                render={({
+                  field: { value: subValue, onChange: subOnChange },
+                }) => (
+                  <RegionDropdown
+                    selectedMainRegion={value}
+                    setSelectedMainRegion={onChange}
+                    selectedSubRegion={subValue}
+                    setSelectedSubRegion={subOnChange}
+                    buttonClassName="w-full flex px-4 py-4 items-center gap-2 self-stretch rounded-xl border focus:border-[rgba(27,141,90,0.93)] focus:outline-none"
+                  />
+                )}
+              />
+            )}
+          />
+          {(errors.mainRegion || errors.subRegion) && (
+            <p className="text-red-500 text-sm mt-1">지역을 선택해주세요</p>
+          )}
+        </div>
+
+        <div>
+          <label className="block text-sm text-text-primary mb-1">금액</label>
+          <Controller
+            name="price"
+            control={control}
+            render={({ field }) => (
+              <Input
+                {...field}
+                type="text"
+                placeholder="₩ 1,000"
+                className="w-full"
+              />
+            )}
+          />
+          {errors.price && (
+            <p className="text-red-500 text-sm mt-1">{errors.price.message}</p>
+          )}
+        </div>
+
+        <div>
+          <label className="block text-sm text-text-primary mb-1">
+            상세 내용
+          </label>
+          <Controller
+            name="contents"
+            control={control}
+            render={({ field }) => (
+              <textarea
+                {...field}
+                placeholder="내용을 입력해주세요"
+                className="resize-none flex w-full h-[13rem] px-4 py-4 items-center gap-2 self-stretch rounded-xl border focus:border-[rgba(27,141,90,0.93)] focus:outline-none"
+              />
+            )}
+          />
+          {errors.contents && (
+            <p className="text-red-500 text-sm mt-1">
+              {errors.contents.message}
+            </p>
+          )}
+        </div>
+
+        <div>
+          <label className="block text-sm font-medium text-text-primary mb-2">
+            사진 첨부
+          </label>
+          <div className="flex gap-4 flex-wrap">
+            {/* 사진 파일 선택 버튼 */}
+            <div
+              className="w-[100px] h-[100px] border-2 border-dashed border-gray-300 rounded-lg flex items-center justify-center cursor-pointer hover:border-gray-400 transition-colors"
+              onClick={handleOpenFileDialog}
+            >
+              <Image
+                src="/icons/camera.png"
+                alt="Upload Image"
+                width={40}
+                height={40}
+              />
+            </div>
+
+            {/* 기존 이미지 썸네일 */}
+            {isEdit &&
+              existingImages.map((image) => (
+                <div
+                  key={image.image_id}
+                  className="w-[100px] h-[100px] relative group"
+                >
+                  <Image
+                    src={image.image_url}
+                    alt={`existing-image-${image.image_id}`}
+                    fill
+                    className="object-cover rounded-lg"
+                  />
+                  <button
+                    type="button"
+                    onClick={() => handleDeleteExistingImage(image.image_id)}
+                    className="absolute top-1 right-1 w-6 h-6 bg-red-500 rounded-full text-white opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center"
+                  >
+                    ×
+                  </button>
+                </div>
+              ))}
+
+            {/* 새로 추가된 이미지 썸네일 */}
+            {watch("newImages")?.map((image, index) => (
+              <div key={index} className="w-[100px] h-[100px] relative group">
+                <Image
+                  src={URL.createObjectURL(image)}
+                  alt={`new-image-${index}`}
+                  fill
+                  className="object-cover rounded-lg"
+                />
+                <button
+                  type="button"
+                  onClick={() => {
+                    const newImages = [...(watch("newImages") || [])];
+                    newImages.splice(index, 1);
+                    setValue("newImages", newImages);
+                  }}
+                  className="absolute top-1 right-1 w-6 h-6 bg-red-500 rounded-full text-white opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center"
+                >
+                  ×
+                </button>
+              </div>
+            ))}
+          </div>
+          <input
+            ref={fileInputRef}
+            type="file"
+            accept="image/*"
+            multiple
+            onChange={handleFileChange}
+            className="hidden"
+          />
+        </div>
+
+        <div className="w-full">
+          <Button
+            type="submit"
+            design="design1"
+            width="full"
+            className="h-[3.5rem]"
+          >
+            {isEdit ? "수정하기" : "등록하기"}
+          </Button>
+        </div>
+      </form>
+    </div>
+  );
+};
+
+export default JobBoardNew;
